added error handling for directory creation and simplified logic handling

+++ pyalgogem/data/data_gathering.py
@@ -12,41 +12,44 @@ import shutil
 
 
 def initialize_data_directory(path=None, folder=None, over_write=False):
-    # Function to create local directory
-    # to store data files for prices
-    #
-    # Returns directory of folder to use
+    """
+    Function to create local directory for historical data storage
 
-    # determine path of folder to use
+    path: string
+        current dir will be used if no path chosen
+    folder: string
+        name of "datasets" will be used for folder if no name chosen
+    over_write: bool
+        delete and over-write if dir already exists
+    return:
+        directory of folder to use
+    """
     if path is None:
         dir_path = os.getcwd()
     else:
         dir_path = str(path)
 
-    # determine folder name
     if folder is None:
         dir_name = 'datasets'
     else:
         dir_name = str(folder)
 
-    # determine spacing based on OS
     if 'win' in sys.platform.lower():
         spacing = '\\'
     else:
         spacing = '/'
     dir = dir_path + spacing + dir_name
 
-    # check if folder to use exists
-    # and overwrite if asked to
-    if os.path.exists(dir):
-        if over_write:
-            shutil.rmtree(dir)
-            os.makedirs(dir)
-    else:
+    if os.path.exists(dir) and over_write:
         try:
+            shutil.rmtree(dir)
             os.makedirs(dir)
         except OSError:
             print("Error creating directory.")
-
+    elif not os.path.exists(dir):
+        try:
+            os.makedirs(dir)
+        except OSError:
+            print("Error creating directory")
     # return location of data directory
     return dir
